feat: user API 정리 및 mutation 적용
